[Groups] moved related group files to Group module and deleted admin group page
