+++ metaspace/webapp/src/modules/Group/GroupsListItem.scss
