+++ metaspace/webapp/src/modules/Group/GroupsListPage.scss
