+++ metaspace/webapp/src/modules/Group/ViewGroupPage.spec.ts
